moved debugs into own .cpp file
